cockroachkvs: export RandomKVs
Export the utility used to generate random CRDB KVs.
